:white_check_mark: Move Word to separate file to increase modularity, add tests, line coverage in dictionary_creator.py: 96%
