+++ dictionary_creator.py
@@ -2,7 +2,7 @@ import math
 import os
 import re
 import subprocess
-from collections import defaultdict, Counter
+from collections import defaultdict
 from datetime import datetime
 
 import dill
@@ -23,74 +23,10 @@ from tokenizers.pre_tokenizers import Whitespace
 from tokenizers.trainers import BpeTrainer
 from tqdm import tqdm
 
+from word import Word
+
 
 class DictionaryCreator(object):
-    class Word(object):
-        def __init__(self, text, lang, qids=None, occurrences_in_bible=0):
-            self.text = text
-            self.iso_language = lang
-            assert (qids is None or type(qids) == set)
-            self.qids = set() if qids is None else qids
-            self.occurrences_in_bible = occurrences_in_bible
-            self.display_text = text
-            self._aligned_words = Counter()
-
-        def __eq__(self, other):
-            if isinstance(other, DictionaryCreator.Word):
-                return self.text == other.text \
-                       and self.iso_language == other.iso_language \
-                       and self.qids == other.qids \
-                       and self.occurrences_in_bible == other.occurrences_in_bible \
-                       and self.display_text == other.display_text \
-                       and self._aligned_words == other._aligned_words
-            return NotImplemented
-
-        def __hash__(self):
-            return hash((self.iso_language, self.text))
-
-        def __str__(self):
-            return f'{self.iso_language}: {self.text}'
-
-        def __repr__(self):
-            return f'{self.iso_language}: {self.text}'
-
-        def get_aligned_words_and_counts(self, words_by_text_by_lang):
-            for word_str, count in self._aligned_words.items():
-                lang, wtxt = word_str.split(': ')
-                word = words_by_text_by_lang[lang][wtxt]
-                yield word, count
-
-        def add_aligned_word(self, word, count=1):
-            # caution: this is not symmetric, todo: make this symmetric
-            self._aligned_words[str(word)] += count
-
-        def remove_alignment(self, word):
-            del self._aligned_words[str(word)]
-
-        def _update_aligned_words(self, lang, removed_word, words_by_text_by_lang):
-            # replace references to a removed word (e.g., drink, drank, drunk)
-            # with references to a merged word (e.g., DRINK)
-            for aligned_word, count in words_by_text_by_lang[lang][removed_word.text]. \
-                    get_aligned_words_and_counts(words_by_text_by_lang):
-                aligned_word.remove_alignment(removed_word)
-                aligned_word.add_aligned_word(self, count)
-            del words_by_text_by_lang[lang][removed_word.text]
-
-        def merge_words(self, words, words_by_text_by_lang, strength_by_lang_by_wtxt_by_lang, changed_variables):
-            # todo: refactor this method by creating a new node instead of modifying an existing one --> call _update_aligned_words only once
-            self.display_text = f'{self.text.upper()} ({len(words) + 1})'
-            for word in words:
-                self.qids.update(
-                    word.qids)  # todo: weight qids by occurrences in Bible when adding semdoms as nodes to graph
-                self.occurrences_in_bible += word.occurrences_in_bible
-                self._aligned_words += word._aligned_words
-                self._update_aligned_words(word.iso_language, word, words_by_text_by_lang)
-                strength_by_lang_by_wtxt_by_lang[word.iso_language].pop(word.text, None)  # remove cache entry
-            self._update_aligned_words(self.iso_language, self, words_by_text_by_lang)
-            strength_by_lang_by_wtxt_by_lang[self.iso_language].pop(self.text, None)  # remove cache entry
-            changed_variables.add('words_by_text_by_lang')
-            changed_variables.add('strength_by_lang_by_wtxt_by_lang')
-
     BIBLES_BY_BID = {
         'bid-eng-asvbt': 'eng-engasvbt.txt',
         'bid-eng-asv': 'eng-eng-asv.txt',
@@ -347,11 +283,12 @@ class DictionaryCreator(object):
             if os.path.isfile(sd_path):
                 self.sds_by_lang[lang] = pd.read_csv(sd_path)
             else:
-                print(f'WARNING: unable to load {sd_path}')
+                print(f'WARNING: Unable to load {sd_path}')
                 # create empty dataframe
                 self.sds_by_lang[lang] = pd.DataFrame(
                     {'cid': [], 'category': [], 'question_index': [], 'question': [], 'answer': []})
-                assert (lang in ('deu', 'rus', 'vie', 'tpi'))
+                if lang not in ('deu', 'rus', 'vie', 'tpi'):
+                    raise FileNotFoundError(f'Unable to load {sd_path}')
             self.changed_variables.add('sds_by_lang')
 
         for bid in tqdm(self.bids, desc='Loading bibles', total=len(self.bids)):
@@ -384,7 +321,7 @@ class DictionaryCreator(object):
                 wtxts = [wtxt.strip().lower() for wtxt in answer.split(',') if wtxt]
                 if lang == 'eng':
                     wtxts = self._lemmatize_english_verse(wtxts)
-                words = {wtxt: self.Word(wtxt.strip(), lang, {qid}) for wtxt in wtxts}
+                words = {wtxt: Word(wtxt.strip(), lang, {qid}) for wtxt in wtxts}
 
                 # add new words to words_by_text_by_lang
                 for word in words.values():
@@ -456,7 +393,7 @@ class DictionaryCreator(object):
                 if wtxt in self.words_by_text_by_lang[lang]:
                     self.words_by_text_by_lang[lang][wtxt].occurrences_in_bible += 1
                 else:
-                    self.words_by_text_by_lang[lang][wtxt] = self.Word(wtxt, lang, set(), 1)
+                    self.words_by_text_by_lang[lang][wtxt] = Word(wtxt, lang, set(), 1)
                 self.changed_variables.add('words_by_text_by_lang')
 
     def _combine_alignments(self):
@@ -1173,7 +1110,7 @@ class DictionaryCreator(object):
     def create_dictionary(self, load=False, save=False, plot_word_lang='eng', plot_word='drink',
                           prediction_method='link prediction'):
         if prediction_method not in ('link prediction', 'tfidf'):
-            raise NotImplementedError
+            raise NotImplementedError(f'Prediction method {prediction_method} not implemented.')
 
         self.preprocess_data(load=load, save=save)
         self.map_words_to_qids(load=load, save=save)
